+++ Reconctruction/Reconstruction/integral.h
@@ -6,7 +6,7 @@
 
 using namespace std;
 
-double Get_integral(const vector<int> dava_v, const double baseline, const int time_scale, const double time_from, const double time_to);
+double Get_integral(const vector<double> dava_v, const double baseline, const int time_scale, const double time_from, const double time_to);
 
 
 #endif // INTEGRAL_H
